Add new attechment functionality to new models
- Giving the ol' refactor tractor a fresh coat of paint
