+++ InvenTree/order/templates/order/po_attachments.html
@@ -21,8 +21,9 @@
 <table class='table table-striped table-condensed' data-toolbar='#attachment-buttons' id='attachment-table'>
     <thead>
         <tr>
-            <th data-field='file' data-searchable='true'>{% trans "File" %}</th>
-            <th data-field='comment' data-searchable='true'>{% trans "Comment" %}</th>
+            <th data-field='file' data-sortable='true' data-searchable='true'>{% trans "File" %}</th>
+            <th data-field='comment' data-sortable='true' data-searchable='true'>{% trans "Comment" %}</th>
+            <th data-field='user' data-sortable='true' data-searchable='true'>{% trans "Uploaded" %}</th>
             <th></th>
         </tr>
     </thead>
@@ -31,6 +32,10 @@
         <tr>
             <td><a href='/media/{{ attachment.attachment }}'>{{ attachment.basename }}</a></td>
             <td>{{ attachment.comment }}</td>
+            <td>
+                {% if attachment.upload_date %}{{ attachment.upload_date }}{% endif %}
+                {% if attachment.user %}<span class='badge'>{{ attachment.user.username }}</div>{% endif %}
+            </td>
             <td>
                 <div class='btn-group' style='float: right;'>    
                     <button type='button' class='btn btn-default btn-glyph attachment-edit-button' url="{% url 'po-attachment-edit' attachment.id %}" data-toggle='tooltip' title='{% trans "Edit attachment" %}'>
