+++ InvenTree/part/views.py
@@ -74,6 +74,11 @@ class PartAttachmentCreate(AjaxCreateView):
     ajax_form_title = _("Add part attachment")
     ajax_template_name = "modal_form.html"
 
+    def post_save(self):
+        """ Record the user that uploaded the attachment """
+        self.object.user = self.request.user
+        self.object.save()
+
     def get_data(self):
         return {
             'success': _('Added attachment')
