+++ InvenTree/stock/templates/stock/item_attachments.html
@@ -20,9 +20,9 @@
 <table class='table table-striped table-condensed' data-toolbar='#attachment-buttons' id='attachment-table'>
     <thead>
         <tr>
-            <th data-field='file' data-searchable='true'>{% trans "File" %}</th>
-            <th data-field='comment' data-searchable='true'>{% trans "Comment" %}</th>
-            <th data-field='user' data-searchable='true'>{% trans "Uploaded" %}</th>
+            <th data-field='file' data-sortable='true' data-searchable='true'>{% trans "File" %}</th>
+            <th data-field='comment' data-sortable='true' data-searchable='true'>{% trans "Comment" %}</th>
+            <th data-field='user' data-sortable='true' data-searchable='true'>{% trans "Uploaded" %}</th>
             <th></th>
         </tr>
     </thead>
